+++ changelog.md
@@ -1,3 +1,10 @@
+# 1.12.0
+- Updated `stdout` api for new `stdout.clearProcessOut` function.
+- Removed space replacing for module names system. (`"chack for updates" != "check-for-updates"`)
+- Added `checkForUpdates` command with `modules/check-for-updates.js` module.
+- Edited `src/load.sh` file.
+- Added focus blocking system to elements inside the `#std-out` element.
+
 # 1.11.0
 - Added "`enable_terminal_like_copy_paste`" setting system for `manifest.json`.